[llm_bench] Add hooks for transformers v5
